make earendil both a lib and bin
