+++ Cargo.toml
@@ -6,6 +6,15 @@ name = "earendil"
 version = "0.0.1"
 edition = "2021"
 
+[lib]
+name = "earendil"
+path = "src/lib.rs"
+
+[[bin]]
+name = "earendil"
+path = "src/main.rs"
+
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
