+++ src/control_protocol.rs
@@ -1,5 +1,5 @@
+use crate::commands::ControlCommands;
 use crate::sockets::socket::Endpoint;
-use crate::ControlCommands;
 use crate::{daemon::ControlProtErr, havens::haven::HavenLocator};
 use anyhow::Context;
 use async_trait::async_trait;
