+++ src/sockets/haven_socket.rs
@@ -1,4 +1,4 @@
-use std::{sync::Arc, time::Duration};
+use std::time::Duration;
 
 use bytes::Bytes;
 use earendil_crypt::{Fingerprint, IdentitySecret};
