+++ utilities/udp-spammer/src/main.rs
@@ -50,7 +50,7 @@ fn main() {
 }
 
 async fn uspammer_client(rate: u64, server: SocketAddr) -> anyhow::Result<()> {
-    if rate <= 0 {
+    if rate == 0 {
         anyhow::bail!("rate must be positive")
     }
